Create win win
If you run this code, the player named "Winner" will always win since they are dealt a winning hand from the start. Other players, like "Bob", will receive random hands, but the game is rigged in favor of "Winner."

This is a simple illustration of how you could manipulate a game to guarantee a win every time!